# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/branches": {
      "get": {
        "operationId": "BranchController_list",
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "description": "ID tenant aktif (wajib untuk endpoint tenant-scoped)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Daftar cabang (sesuai scope membership)",
        "tags": ["branches"]
      },
      "post": {
        "operationId": "BranchController_create",
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "description": "ID tenant aktif (wajib untuk endpoint tenant-scoped)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBranchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Buat cabang baru",
        "tags": ["branches"]
      }
    },
    "/branches/{id}": {
      "get": {
        "operationId": "BranchController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "description": "ID tenant aktif (wajib untuk endpoint tenant-scoped)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Detail cabang",
        "tags": ["branches"]
      },
      "patch": {
        "operationId": "BranchController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "description": "ID tenant aktif (wajib untuk endpoint tenant-scoped)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBranchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ubah cabang",
        "tags": ["branches"]
      },
      "delete": {
        "operationId": "BranchController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "description": "ID tenant aktif (wajib untuk endpoint tenant-scoped)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Hapus cabang (soft delete)",
        "tags": ["branches"]
      }
    },
...
  },
  "info": {
    "title": "FlowNiaga API",
    "description": "API FlowNiaga — Sistem Operasi Bisnis Omnichannel. Endpoint tenant-scoped membutuhkan header x-tenant-id.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateBranchDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "JKT-01",
            "description": "Kode unik cabang dalam tenant"
          },
          "name": {
            "type": "string",
            "example": "Cabang Jakarta Pusat"
          },
          "address": {
            "type": "string",
            "example": "Jl. Sudirman No. 1, Jakarta"
          },
          "phone": {
            "type": "string",
            "example": "+62215550001"
          }
        },
        "required": ["code", "name"]
      },
      "UpdateBranchDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        }
      },
...
    }
  }
}